Refactored code and updated Ollama models API URL
- Created a new file `LlmServerConverter.cs` containing the `LlmServerConverter` class, which was previously in `EntwineLLMOptions.cs`.
- Updated the API URL in `OllamaServer.cs` from `api/list_models` to `api/tags`.
