+++ src/EntwineLLM/LlmServerConverter.cs
@@ -1,0 +1,37 @@
+﻿using System;
+using System.ComponentModel;
+using System.Globalization;
+using System.Linq;
+using System.Reflection;
+
+namespace EntwineLlm.Models
+{
+    public class LlmServerConverter : TypeConverter
+        {
+            private static readonly LLMServerBase[] Servers = [.. Assembly.GetExecutingAssembly().GetTypes()
+                .Where(type => type.IsSubclassOf(typeof(LLMServerBase)) && !type.IsAbstract)
+                .Select(type => (LLMServerBase)Activator.CreateInstance(type))];
+
+            public override bool GetStandardValuesSupported(ITypeDescriptorContext context) => true;
+
+            public override bool GetStandardValuesExclusive(ITypeDescriptorContext context) => true;
+
+            public override StandardValuesCollection GetStandardValues(ITypeDescriptorContext context) => new(Servers);
+
+            public override bool CanConvertTo(ITypeDescriptorContext context, Type destinationType)
+                => destinationType == typeof(string);
+
+            public override object ConvertTo(ITypeDescriptorContext context, CultureInfo culture, object value, Type destinationType)
+                => destinationType == typeof(string)
+                ? ((LLMServerBase)value).Name
+                : base.ConvertTo(context, culture, value, destinationType);
+
+            public override bool CanConvertFrom(ITypeDescriptorContext context, Type sourceType)
+                => sourceType == typeof(string);
+
+            public override object ConvertFrom(ITypeDescriptorContext context, CultureInfo culture, object value)
+                => value is string stringValue
+                ? Servers.FirstOrDefault(server => server.Name == stringValue)
+                : base.ConvertFrom(context, culture, value);
+        }
+}
